refactor(layout): Change type to height (#185)
- Change `type` to `height` parameter to make it more descriptive
- Renamed:
  - `type="stretch"` to `height="auto"`
  - `type="auto"` to `height="fit"` (this was and still is the default)
  - `type="parent"` to `height="parent"` (no change)

The reason for the change is to make the API more consistent. `"fit"` and `"auto"`
are valid values for Grid `size` with very specific meaning. We should use the same
terminology to identify the same concepts. This is similar to the prior change from
`align="middle"` to `align="center"` for consistency with `justify`.

TODO: Update wiki (after this PR is merged)

BREAKING CHANGE