Add base Component (#196)
Add base component, behaves like Grid but only takes a custom `style` attribute
instead of `margin` and `padding`. Note that `<Grid />` now extends `<Base/>` but
sets the appropriate css for margins and paddings based on gutter size which is
set at `24px`.

Base should really be used directly only as a last resort option. It may be needed as part of styleguide components but should not be needed for layout.