feat(dev): Allow override dev mode toggles
Add boolean override parameter to `toggleOverlay()` and `toggleColor()` to force setting or unsetting (instead of toggling) a mode.

This behavior mimics jQuery or and DOM API's (e.g. `$().toggleClass` and `classList.toggle`) so it should be familiar for consumers.

Additionally, this is important for hot reloading where the setup code may runs multiple times and would otherwise keep toggling colors on and off

Closes https://github.com/obartra/reflex/issues/258